+++ sdk/core/Azure.Core.Experimental/src/RequestOptions.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Collections.Generic;
+using System.Threading;
+using Azure.Core;
+using Azure.Core.Pipeline;
+
+namespace Azure
+{
+    /// <summary>
+    /// Options which can be used to control the behavior of a request sent by a client.
+    /// </summary>
+    public class RequestOptions : RequestContext
+    {
+        /// <summary>
+        /// </summary>
+        private new CancellationToken CancellationToken { get; set; }
+    }
+}